feat: Add files for nextjs vercel deployment
