--- backend/.env.demo
@@ -1,5 +1,0 @@
-JWT_SECRET=somevalueofyourchoice
-DB_HOST =
-DB_USER =
-DB_PASSWORD =
-DB_NAME =
