--- backend/.gitignore
@@ -1,3 +1,0 @@
-node_modules/*
-package-lock.json
-.env
