+++ frontend/package.json
@@ -9,6 +9,12 @@
     "lint": "eslint"
   },
   "dependencies": {
+    "bcrypt": "^6.0.0",
+    "cors": "^2.8.6",
+    "dotenv": "^17.4.2",
+    "express": "^5.2.1",
+    "jsonwebtoken": "^9.0.3",
+    "mysql2": "^3.23.1",
     "next": "16.2.9",
     "react": "19.2.4",
     "react-dom": "19.2.4"
